+++ src/main/java/com/chargebee/Environment.java
@@ -32,7 +32,7 @@ public class Environment {
 
     public static final String API_VERSION = "v2";
     
-    public static final String LIBRARY_VERSION = "2.1.4";
+    public static final String LIBRARY_VERSION = "2.1.5";
 
     private final String apiBaseUrl;